+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # CHANGELOG
 
+## v0.3 - 2023-04-24
+
+- Allies inform you when they're invading a planet
+
 ## v0.2 - 2023-04-20
 
-- Fixed issue with AI allies not telling you where they landing in some circumstances
+- Fixed issue with AI allies not telling you where they landed in some circumstances
 - Support an AI ally with multiple commanders landing on multiple worlds
 - Allies inform you when they're colonising a new planet